Better mod integration methods :P Can now run functions. Config.

+++ src/main/java/tfc/dynamic_weaponary/Utils/Tool/Material.java
@@ -4,18 +4,18 @@ import tfc.dynamic_weaponary.Utils.DrawingUtils;
 
 public class Material {
 	public int durability;
-	public int strength;
-	public float weight;
+	public double strength;
+	public double weight;
 	public int color;
 	
-	public Material(int durability, int strength, float weight, int color) {
+	public Material(int durability, double strength, double weight, int color) {
 		this.durability = durability;
 		this.strength = strength;
 		this.weight = weight;
 		this.color = color;
 	}
 	
-	public Material(int durability, int strength, float weight, DrawingUtils.ColorHelper color) {
+	public Material(int durability, double strength, double weight, DrawingUtils.ColorHelper color) {
 		this.durability = durability;
 		this.strength = strength;
 		this.weight = weight;
